--- lib/firebase_options.dart
@@ -1,83 +1,0 @@
-// File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
-import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
-import 'package:flutter/foundation.dart'
-    show defaultTargetPlatform, kIsWeb, TargetPlatform;
-
-/// Default [FirebaseOptions] for use with your Firebase apps.
-///
-/// Example:
-/// ```dart
-/// import 'firebase_options.dart';
-/// // ...
-/// await Firebase.initializeApp(
-///   options: DefaultFirebaseOptions.currentPlatform,
-/// );
-/// ```
-class DefaultFirebaseOptions {
-  static FirebaseOptions get currentPlatform {
-    if (kIsWeb) {
-      return web;
-    }
-    switch (defaultTargetPlatform) {
-      case TargetPlatform.android:
-        return android;
-      case TargetPlatform.iOS:
-        return ios;
-      case TargetPlatform.macOS:
-        return macos;
-      case TargetPlatform.windows:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for windows - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      case TargetPlatform.linux:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions have not been configured for linux - '
-          'you can reconfigure this by running the FlutterFire CLI again.',
-        );
-      default:
-        throw UnsupportedError(
-          'DefaultFirebaseOptions are not supported for this platform.',
-        );
-    }
-  }
-
-  static const FirebaseOptions web = FirebaseOptions(
-    apiKey: 'AIzaSyDUIoxQRgiOXJek-hlYYJUpDanRYbNaT_M',
-    appId: '1:468136851061:web:c412779484eca6317d2a3c',
-    messagingSenderId: '468136851061',
-    projectId: 'linkstagram-1',
-    authDomain: 'linkstagram-1.firebaseapp.com',
-    storageBucket: 'linkstagram-1.appspot.com',
-    measurementId: 'G-DQHTNNH8BF',
-  );
-
-  static const FirebaseOptions android = FirebaseOptions(
-    apiKey: 'AIzaSyAwDPiyPYVTDF_GDS8CGl4S1CqptR7y6lQ',
-    appId: '1:468136851061:android:e34c0fd4389bd8d37d2a3c',
-    messagingSenderId: '468136851061',
-    projectId: 'linkstagram-1',
-    storageBucket: 'linkstagram-1.appspot.com',
-  );
-
-  static const FirebaseOptions ios = FirebaseOptions(
-    apiKey: 'AIzaSyDogl05yjpkMCbVNMd6-0IA56ztLR2l_vI',
-    appId: '1:468136851061:ios:10556e91d7b4001b7d2a3c',
-    messagingSenderId: '468136851061',
-    projectId: 'linkstagram-1',
-    storageBucket: 'linkstagram-1.appspot.com',
-    iosClientId: '468136851061-r57j5trmsidfsturj3tg7h11d2nei12m.apps.googleusercontent.com',
-    iosBundleId: 'com.example.linkstagram',
-  );
-
-  static const FirebaseOptions macos = FirebaseOptions(
-    apiKey: 'AIzaSyDogl05yjpkMCbVNMd6-0IA56ztLR2l_vI',
-    appId: '1:468136851061:ios:cc1887f5f655ad8d7d2a3c',
-    messagingSenderId: '468136851061',
-    projectId: 'linkstagram-1',
-    storageBucket: 'linkstagram-1.appspot.com',
-    iosClientId: '468136851061-09ht1mqhqgjcsh0t84348jqpovmcu790.apps.googleusercontent.com',
-    iosBundleId: 'com.example.linkstagram.RunnerTests',
-  );
-}
