--- lib/responsive/mobile_screen_layout.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MobileScreenLayout extends StatelessWidget {
-  const MobileScreenLayout({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('this is mobile'),
-      ),
-    );
-  }
-}
