--- lib/responsive/responsive_layout_screen.dart
@@ -1,27 +1,0 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
-import 'package:flutter/material.dart';
-
-import 'package:linkstagram/utils/dimensions.dart';
-
-class ResponsiveLayout extends StatelessWidget {
-  const ResponsiveLayout({
-    Key? key,
-    required this.webScreenLayout,
-    required this.mobileScreenLayout,
-  }) : super(key: key);
-
-  final Widget webScreenLayout;
-  final Widget mobileScreenLayout;
-
-  @override
-  Widget build(BuildContext context) {
-    return LayoutBuilder(
-      builder: (context, constraints) {
-        if (constraints.maxWidth > webScreenSize) {
-          return webScreenLayout;
-        }
-        return mobileScreenLayout;
-      },
-    );
-  }
-}
