--- lib/responsive/web_screen_layout.dart
@@ -1,14 +1,0 @@
-import 'package:flutter/material.dart';
-
-class WebScreenLayout extends StatelessWidget {
-  const WebScreenLayout({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Text('this is web'),
-      ),
-    );
-  }
-}
