--- lib/services/storage_methods.dart
@@ -1,20 +1,0 @@
-import 'dart:typed_data';
-
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_storage/firebase_storage.dart';
-
-class StorageMethods {
-  final FirebaseStorage _storage = FirebaseStorage.instance;
-  final FirebaseAuth _auth = FirebaseAuth.instance;
-
-  Future<String> uploadImageToStorage(
-      String name, Uint8List file, bool isPost) async {
-    Reference ref = _storage.ref().child(name).child(_auth.currentUser!.uid);
-
-    UploadTask uploadTask = ref.putData(file);
-
-    TaskSnapshot snap = await uploadTask;
-    String downloadUrl = await snap.ref.getDownloadURL();
-    return downloadUrl;
-  }
-}
