--- lib/style/app_text_style.dart
@@ -1,1 +1,0 @@
-class AppTextStyle {}
