--- lib/utils/dimensions.dart
@@ -1,3 +1,0 @@
-import 'package:flutter/material.dart';
-
-const webScreenSize = 600;
