--- lib/utils/helper.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-showSnackBar(BuildContext context, String text) {
-  ScaffoldMessenger.of(context).showSnackBar(
-    SnackBar(
-      content: Text(text),
-    ),
-  );
-}
