--- lib/widgets/post_widget.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class PostWidget extends StatelessWidget {
-  const PostWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
